requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.16.0
python-docx>=0.8.11
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
requests>=2.31.0